+++ frontend/src/_colors.scss
@@ -1,0 +1,6 @@
+/* Color Theme Swatches in Hex */
+$Deep-Blue-hex: #053868;
+$Tan-hex: #EDF5E1;
+$Light-Green-hex: #8EE4AF;
+$Green-hex: #5CDB95;
+$Dark-Green-hex: #379683;
