--- frontend/src/bundle.css
@@ -1,4 +1,0 @@
-@import "./0reset.css";
-@import "./components/nav/navbar.scss";
-@import "./app.scss";
-@import "./components/session/session_form.scss"
