+++ frontend/src/components/app.jsx
@@ -6,6 +6,7 @@ import NavBarContainer from "./nav/navbar_container";
 import MainPage from "./main/main_page";
 import LoginFormContainer from "./session/login_form_container";
 import SignupFormContainer from "./session/signup_form_container";
+import '../app.scss'
 
 const App = () => (
   <div>
