+++ frontend/src/components/main/main_page.jsx
